Changed from Summary to Detail
This update reflects what would be necessary for a Kitchen staff/chef to actually plate the dishes ordered. I added the menu options so the options selected by the customer are reflected in the details for each order. I also added buttons for updating the status of the order through the process after Received. Namely buttons for Preparation, Ready for Pickup and Complete.