Cosmetic preference
Added alignment to the cell for the buttons so there it just looks cleaner. I am still planning to update the layout so I can get 3 or 4 orders across a row with buttons at the bottom with a size and spacing to make sure the user does not accidentally tap the wrong button.

Otherwise, looks good. Thank you.