Small tweak
Added label-default to buttons to make their text show white rather than black. Much more readable.
Any your tweaks are great.

Three questions remain: 
1 - how to make the email process kick off and 
2 - is there a way to make the cards have a minimum width so they don't compress past the size of the btn-group? On a phone they compress way too small. (granted, not everyone will be using a phone to process orders in their kitchen, but I have a food truck and money is tight, so we use whatever tech we already own like old tablets, phones, etc.)
3 - have you considered renaming this to anything I suggested?
Not looking for answers on 1 & 2, but for sure highlighting work that still needs to be done to bring this to final production. I will for sure use this in its present state and even more functionality when the email is getting sent. Although, text message would be better (we tend to get better response when we text customers than when we email).

Thank you again for your accepting my tweaks and reviewing my sloppy additions.